Add pluggable authorization support to broker-core
