+++ modules/broker-amqp/src/main/java/io/ballerina/messaging/broker/amqp/codec/frames/ConnectionSecureOk.java
@@ -23,6 +23,7 @@ import io.ballerina.messaging.broker.amqp.codec.BlockingTask;
 import io.ballerina.messaging.broker.amqp.codec.ConnectionException;
 import io.ballerina.messaging.broker.amqp.codec.auth.SaslAuthenticationStrategy;
 import io.ballerina.messaging.broker.amqp.codec.handlers.AmqpConnectionHandler;
+import io.ballerina.messaging.broker.auth.UsernamePrincipal;
 import io.ballerina.messaging.broker.common.data.types.LongString;
 import io.ballerina.messaging.broker.common.data.types.ShortString;
 import io.netty.buffer.ByteBuf;
@@ -32,6 +33,7 @@ import io.netty.util.AttributeKey;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.security.auth.Subject;
 import javax.security.sasl.SaslException;
 import javax.security.sasl.SaslServer;
 
@@ -76,6 +78,9 @@ public class ConnectionSecureOk extends MethodFrame {
                     byte[] challenge = saslServer.evaluateResponse(response.getBytes());
                     if (saslServer.isComplete()) {
                         ctx.writeAndFlush(new ConnectionTune(256, 65535, 0));
+                        Subject subject = new Subject();
+                        subject.getPrincipals().add(new UsernamePrincipal(saslServer.getAuthorizationID()));
+                        connectionHandler.secureBroker(subject);
                     } else {
                         ctx.channel().attr(AttributeKey.valueOf(SaslAuthenticationStrategy.SASL_SERVER_ATTRIBUTE))
                             .set(null);
