+++ modules/broker-core/src/main/java/io/ballerina/messaging/broker/core/DefaultBrokerFactory.java
@@ -1,0 +1,40 @@
+/*
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.ballerina.messaging.broker.core;
+
+import io.ballerina.messaging.broker.common.StartupContext;
+
+import javax.security.auth.Subject;
+
+/**
+ * Implementation to return the {@link BrokerImpl} object
+ */
+public class DefaultBrokerFactory implements BrokerFactory {
+
+    private final Broker broker;
+
+    public DefaultBrokerFactory(StartupContext startupContext) {
+        this.broker = startupContext.getService(Broker.class);
+    }
+
+    @Override
+    public Broker getBroker(Subject subject) {
+        return broker;
+    }
+}
