+++ modules/broker-core/src/main/java/io/ballerina/messaging/broker/core/SecureBrokerFactory.java
@@ -1,0 +1,43 @@
+/*
+ * Copyright (c) 2018, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
+ *
+ * WSO2 Inc. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+package io.ballerina.messaging.broker.core;
+
+import io.ballerina.messaging.broker.auth.AuthManager;
+import io.ballerina.messaging.broker.common.StartupContext;
+
+import javax.security.auth.Subject;
+
+/**
+ * Implementation to return new {@link SecureBrokerImpl} object providing {@link Subject}
+ */
+public class SecureBrokerFactory implements BrokerFactory {
+
+    private final Broker broker;
+    private final AuthManager authManager;
+
+    public SecureBrokerFactory(StartupContext startupContext) {
+        this.broker = startupContext.getService(Broker.class);
+        this.authManager = startupContext.getService(AuthManager.class);
+    }
+
+    @Override
+    public Broker getBroker(Subject subject) {
+        return new SecureBrokerImpl(broker, authManager, subject);
+    }
+}
